[EH] modify : separated passwordverify as middleware
