+++ routes/auth.js
@@ -6,6 +6,7 @@ const DB = require("../database/maria"); // DB 정보 가져오기
 const { smtpTransport } = require("../config/email");
 require("dotenv").config(); // jwt secret key 가져오기
 const { verifyToken } = require("./authmiddleware");
+const { verifypassword } = require("./passmiddleware");
 
 //body-parser
 router.use(express.urlencoded({ extended: true }));
@@ -58,12 +59,12 @@ router.post("/register", async function (req, res) {
       res.status(400).json({ text: "ErrorCode:400, 잘못된 요청입니다." });
     } else {
       console.log("user register success");
-      res.json({ status: "success" });
+      res.status(200).json({ status: "success" });
     }
   });
 });
 
-const emailVerification = async function (req, res) {
+router.post("/email", async function (req, res) {
   const email = req.body.email;
   const number = generateRandom(111111, 999999);
   const mailoptions = {
@@ -97,68 +98,15 @@ const emailVerification = async function (req, res) {
       );
     } else {
       //이미 존재하는 유저.
-      res.json({ status: "email duplicate" });
+      res.status(400).json({
+        status: "email duplicate",
+        text: "이메일이 중복됩니다. 다른 이메일을 입력해주세요",
+      });
     }
   } catch (e) {
     console.log(e);
     res.status(400).json({ text: "ErrorCode:400, 잘못된 요청입니다." });
   }
-};
-
-//비밀번호 일치 확인 미들웨어
-const passwordverify = async function (req, res, next) {
-  const email = req.body.email;
-  const pw = req.body.password;
-  try {
-    const [data] = await DB.promise().query(
-      `select salt, password from users where email = ?`,
-      email
-    );
-    const db_salt = data[0].salt;
-    const db_pw = data[0].password;
-    const req_pw = [
-      crypto.pbkdf2Sync(pw, db_salt, 9999, 64, "sha512").toString("base64"),
-    ];
-    if (req_pw == db_pw) {
-      console.log("비밀번호 일치");
-      next();
-    } else {
-      console.log("비밀번호가 일치하지 않습니다.");
-      res.json({ status: "fail" });
-    }
-  } catch (e) {
-    console.log("Password verify error");
-    res.status(400).json({ text: "ErrorCode:400, 잘못된 요청입니다." });
-  }
-};
-
-// email 중복 여부 확인
-router.get("/verifyid", (req, res) => {
-  const req_email = req.body.email;
-  DB.query(
-    `select email from users where email = ?`,
-    req_email,
-    (err, result, fields) => {
-      if (err) {
-        console.log("유저 이메일이 없습니다.");
-        res.status(400).json({ text: "ErrorCode:400, 잘못된 요청입니다." });
-      } else {
-        if (result[0] == undefined) {
-          console.log("이 이메일은 사용 가능합니다.");
-          res.json({ status: "success" });
-        } else {
-          const idverify = result[0].email;
-          if (req_email == idverify) {
-            console.log("이메일이 중복됩니다. 다른 이메일을 입력해주세요");
-            res.status(400).json({ text: "ErrorCode:400, 잘못된 요청입니다." });
-          } else {
-            console.log("이 이메일은 사용 가능합니다.");
-            res.json({ status: "success" });
-          }
-        }
-      }
-    }
-  );
 });
 
 /*  
@@ -168,7 +116,7 @@ router.get("/verifyid", (req, res) => {
     토큰을 받게 되면 그 userid로 필요한 정보를 받는다.
 */
 
-router.post("/login", passwordverify, async (req, res) => {
+router.post("/login", verifypassword, async (req, res) => {
   const { email, password } = req.body;
   try {
     const [data] = await DB.promise().query(
@@ -189,7 +137,7 @@ router.post("/login", passwordverify, async (req, res) => {
     );
 
     console.log("Login success");
-    res.json({
+    res.status(200).json({
       code: 200,
       message: "토큰이 발급되었습니다.",
       token: token,
@@ -212,5 +160,4 @@ router.get("/token-test", verifyToken, (req, res) => {
   });
 });
 
-router.post("/email", emailVerification);
 module.exports = router;
